# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VERIF-AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateStudentProfile
      tags: [student]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileInput"
      responses:
        "200":
          description: Updated student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
  /recruiter/profile:
    get:
      operationId: getRecruiterProfile
      tags: [recruiter]
      summary: Get recruiter profile
      responses:
        "200":
          description: Recruiter profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruiterProfile"
    put:
      operationId: updateRecruiterProfile
      tags: [recruiter]
      summary: Update recruiter profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecruiterProfileInput"
      responses:
        "200":
          description: Updated recruiter profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruiterProfile"
...
  /recruiter/student/{studentId}:
    get:
      operationId: getStudentForRecruiter
      tags: [recruiter]
      summary: Get a student profile (for recruiter view)
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /recruiter/applications/{applicationId}:
    patch:
      operationId: updateApplicationStatus
      tags: [recruiter]
      summary: Update application status
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusInput"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── AI ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, description, company, location, jobType, skills, recruiterId, postedAt, isActive, applicantCount]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        company: { type: string }
        location: { type: string }
        jobType: { type: string }
        salary: { type: ["string", "null"] }
        skills: { type: array, items: { type: string } }
        recruiterId: { type: integer }
        postedAt: { type: string }
        isActive: { type: boolean }
        applicantCount: { type: integer }
...
    Application:
      type: object
      required: [id, jobId, studentId, status, appliedAt]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        studentId: { type: integer }
        status: { type: string }
        appliedAt: { type: string }
        job:
          $ref: "#/components/schemas/Job"
...
    ApplicationStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string }

    # ── STUDENT ─────────────────────────────────────────────────────────────
    StudentProfile:
      type: object
      required: [id, displayName, email, role, createdAt]
      properties:
        id: { type: integer }
        displayName: { type: string }
        email: { type: string }
        role: { type: string }
        college: { type: ["string", "null"] }
        skills: { type: array, items: { type: string } }
        linkedinUrl: { type: ["string", "null"] }
        githubUrl: { type: ["string", "null"] }
        portfolioUrl: { type: ["string", "null"] }
        resumeUrl: { type: ["string", "null"] }
        resumeFilename: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        appliedJobIds: { type: array, items: { type: integer } }
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        createdAt: { type: string }
    StudentProfileInput:
      type: object
      properties:
        displayName: { type: string }
        college: { type: ["string", "null"] }
        skills: { type: array, items: { type: string } }
        linkedinUrl: { type: ["string", "null"] }
        githubUrl: { type: ["string", "null"] }
        portfolioUrl: { type: ["string", "null"] }
...
    Certificate:
      type: object
      required: [id, filename, url, uploadedAt]
      properties:
        id: { type: integer }
        filename: { type: string }
        url: { type: string }
        uploadedAt: { type: string }

    # ── RECRUITER ───────────────────────────────────────────────────────────
    RecruiterProfile:
      type: object
      required: [id, displayName, email, role, createdAt]
      properties:
        id: { type: integer }
        displayName: { type: string }
        email: { type: string }
        role: { type: string }
        companyName: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    RecruiterProfileInput:
      type: object
      properties:
        displayName: { type: string }
        companyName: { type: ["string", "null"] }
...