# lib/api-zod/src/generated/types/recruiterProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERIF-AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecruiterProfile {
  id: number;
  displayName: string;
  email: string;
  role: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}