# lib/api-zod/src/generated/types/recruiterProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERIF-AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecruiterProfileInput {
  displayName?: string;
  /** @nullable */
  companyName?: string | null;
}