# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERIF-AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Certificate } from './certificate';

export interface StudentProfile {
  id: number;
  displayName: string;
  email: string;
  role: string;
  /** @nullable */
  college?: string | null;
  skills?: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  resumeFilename?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  appliedJobIds?: number[];
  certificates?: Certificate[];
  createdAt: string;
}