# lib/api-zod/src/generated/types/studentProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERIF-AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfileInput {
  displayName?: string;
  /** @nullable */
  college?: string | null;
  skills?: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
}